pyproject.toml
[project]
name = "migration"
version = "0.1.0"
description = "A tool to migrate from Marzban to Marzneshin in three simple steps"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "apscheduler>=3.11.0",
    "fastapi-responses>=0.2.1",
    "fastapi>=0.115.5",
    "httpx>=0.28.0",
    "pydantic>=2.10.2",
    "pymysql==1.1.1",
    "python-decouple>=3.8",
    "pytz==2024.1",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.1",
]
